Introduce enum to make date array usage clearer
An enum has been introduced in the main.cpp file to make the usage of date array more understandable. Earlier, the date array was directly being accessed using indices such as 0, 1, 2 which was hard to understand. Now, these indices have been replaced by 'month', 'day', 'year' respectively, which makes the code easier to comprehend. Also, the calculation of 'months_here' variable and 'levelProgress' variable now includes the current month. This change enhances the readability and maintainability of the code.